requirements.txt
alembic==1.13.3
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
greenlet==3.1.1
Jinja2==3.1.4
kiwisolver==1.4.7
loguru==0.7.2
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
optuna==4.0.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==74.1.2
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.2
torch==2.4.1
torch-summary==1.4.5
torchvision==0.19.1
tqdm==4.66.6
triton==3.0.0
typing_extensions==4.12.2
